// src/api/prefect.tsx
export interface components {
    schemas: {
// ...
        BlockDocument: {
            /**
             * Id
             * Format: uuid
             */
            id?: string;
            /** Created */
            created?: string | null;
            /** Updated */
            updated?: string | null;
            /**
             * Name
             * @description The block document's name. Not required for anonymous block documents.
             */
            name?: string | null;
            /**
             * Data
             * @description The block document's data
             */
            data?: Record<string, never>;
            /**
             * Block Schema Id
             * Format: uuid
             * @description A block schema ID
             */
            block_schema_id: string;
            /** @description The associated block schema */
            block_schema?: components["schemas"]["BlockSchema"] | null;
            /**
             * Block Type Id
             * Format: uuid
             * @description A block type ID
             */
            block_type_id: string;
            /**
             * Block Type Name
             * @description The associated block type's name
             */
            block_type_name?: string | null;
            /** @description The associated block type */
            block_type?: components["schemas"]["BlockType"] | null;
            /**
             * Block Document References
             * @description Record of the block document's references
             */
            block_document_references?: {
                [key: string]: Record<string, never>;
            };
            /**
             * Is Anonymous
             * @description Whether the block is anonymous (anonymous blocks are usually created by Prefect automatically)
             * @default false
             */
            is_anonymous: boolean;
        };
        /**
         * BlockDocumentCreate
         * @description Data used by the Prefect REST API to create a block document.
         */
        BlockDocumentCreate: {
            /**
             * Name
             * @description The block document's name. Not required for anonymous block documents.
             */
            name?: string | null;
            /**
             * Data
             * @description The block document's data
             */
            data?: Record<string, never>;
            /**
             * Block Schema Id
             * Format: uuid
             * @description A block schema ID
             */
            block_schema_id: string;
            /**
             * Block Type Id
             * Format: uuid
             * @description A block type ID
             */
            block_type_id: string;
            /**
             * Is Anonymous
             * @description Whether the block is anonymous (anonymous blocks are usually created by Prefect automatically)
             * @default false
             */
            is_anonymous: boolean;
        };
// ...
        BlockDocumentUpdate: {
            /**
             * Block Schema Id
             * @description A block schema ID
             */
            block_schema_id?: string | null;
            /**
             * Data
             * @description The block document's data
             */
            data?: Record<string, never>;
            /**
             * Merge Existing Data
             * @default true
             */
            merge_existing_data: boolean;
        };
// ...
        Body_read_block_documents_block_documents_filter_post: {
            block_documents?: components["schemas"]["BlockDocumentFilter"] | null;
            block_types?: components["schemas"]["BlockTypeFilter"] | null;
            block_schemas?: components["schemas"]["BlockSchemaFilter"] | null;
            /**
             * Include Secrets
             * @description Whether to include sensitive values in the block document.
             * @default false
             */
            include_secrets: boolean;
            /** @default NAME_ASC */
            sort: components["schemas"]["BlockDocumentSort"] | null;
            /**
             * Offset
             * @default 0
             */
            offset: number;
            /**
             * Limit
             * @description Defaults to PREFECT_API_DEFAULT_LIMIT if not provided.
             */
            limit?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}